Moving models to correct locations.
