+++ stac_fastapi/extensions/stac_fastapi/extensions/core/transaction.py
@@ -1,52 +1,20 @@
 """Transaction extension."""
 
-from typing import Any, Dict, List, Literal, Optional, Type, Union
+from typing import List, Literal, Optional, Type, Union
 
 import attr
 from fastapi import APIRouter, Body, FastAPI, Header
 from stac_pydantic import Collection, Item, ItemCollection
-from stac_pydantic.shared import BBox, MimeTypes
+from stac_pydantic.shared import MimeTypes
 from starlette.responses import JSONResponse, Response
-from typing_extensions import Annotated, TypedDict
+from typing_extensions import Annotated
 
 from stac_fastapi.api.models import CollectionUri, ItemUri
 from stac_fastapi.api.routes import create_async_endpoint
 from stac_fastapi.types.config import ApiSettings
 from stac_fastapi.types.core import AsyncBaseTransactionsClient, BaseTransactionsClient
 from stac_fastapi.types.extension import ApiExtension
-
-
-class PartialCollection(TypedDict, total=False):
-    """Partial STAC Collection."""
-
-    type: Optional[str]
-    stac_version: Optional[str]
-    stac_extensions: Optional[List[str]]
-    id: Optional[str]
-    title: Optional[str]
-    description: Optional[str]
-    links: List[Dict[str, Any]]
-    keywords: Optional[List[str]]
-    license: Optional[str]
-    providers: Optional[List[Dict[str, Any]]]
-    extent: Optional[Dict[str, Any]]
-    summaries: Optional[Dict[str, Any]]
-    assets: Optional[Dict[str, Any]]
-
-
-class PartialItem(TypedDict, total=False):
-    """Partial STAC Item."""
-
-    type: Optional[Literal["Feature"]]
-    stac_version: Optional[str]
-    stac_extensions: Optional[List[str]]
-    id: Optional[str]
-    geometry: Optional[Dict[str, Any]]
-    bbox: Optional[BBox]
-    properties: Optional[Dict[str, Any]]
-    links: Optional[List[Dict[str, Any]]]
-    assets: Optional[Dict[str, Any]]
-    collection: Optional[str]
+from stac_fastapi.types.stac import PartialCollection, PartialItem, PatchOperation
 
 
 @attr.s
@@ -54,39 +22,6 @@ class PostItem(CollectionUri):
     """Create Item."""
 
     item: Annotated[Union[Item, ItemCollection], Body()] = attr.ib(default=None)
-
-
-@attr.s
-class PatchAddReplaceTest:
-    """Add, Replace or Test Operation."""
-
-    path: str = attr.ib()
-    op: Literal["add", "replace", "test"] = attr.ib()
-    value: Any = attr.ib()
-
-
-@attr.s
-class PatchRemove:
-    """Remove Operation."""
-
-    path: str = attr.ib()
-    op: Literal["remove"] = attr.ib()
-
-
-@attr.s
-class PatchMoveCopy:
-    """Move or Copy Operation."""
-
-    path: str = attr.ib()
-    op: Literal["move", "copy"] = attr.ib()
-
-    def __attrs_init__(self, *args, **kwargs):
-        """Init function to add 'from' field."""
-        super().__init__(*args, **kwargs)
-        self.__setattr__("from", kwargs["from"])
-
-
-PatchOperation = Union[PatchAddReplaceTest, PatchMoveCopy, PatchRemove]
 
 
 @attr.s
