+++ stac_fastapi/extensions/tests/test_transaction.py
@@ -11,7 +11,7 @@ from stac_fastapi.api.app import StacApi
 from stac_fastapi.extensions.core import TransactionExtension
 from stac_fastapi.types.config import ApiSettings
 from stac_fastapi.types.core import BaseCoreClient, BaseTransactionsClient
-from stac_fastapi.types.transaction import PatchOperation
+from stac_fastapi.types.stac import PatchOperation
 
 
 class DummyCoreClient(BaseCoreClient):
@@ -145,7 +145,9 @@ def test_merge_patch_item(client: TestClient, item: Item) -> None:
 
 
 def test_json_patch_item(client: TestClient) -> None:
-    operations = [{"op": "add", "path": "properties.new_prop", "value": "new_prop_value"}]
+    operations = [
+        {"op": "add", "path": "properties.new_prop", "value": "new_prop_value"}
+    ]
     headers = {"Content-Type": "application/json-patch+json"}
     response = client.patch(
         "/collections/a-collection/items/an-item",
@@ -189,7 +191,9 @@ def test_merge_patch_collection(client: TestClient, collection: Collection) -> N
 
 
 def test_json_patch_collection(client: TestClient) -> None:
-    operations = [{"op": "add", "path": "summaries.new_prop", "value": "new_prop_value"}]
+    operations = [
+        {"op": "add", "path": "summaries.new_prop", "value": "new_prop_value"}
+    ]
     headers = {"Content-Type": "application/json-patch+json"}
     response = client.patch(
         "/collections/a-collection/items/an-item",
@@ -269,7 +273,9 @@ def collection() -> Collection:
         "description": "A test collection",
         "extent": {
             "spatial": {"bbox": [[-180, -90, 180, 90]]},
-            "temporal": {"interval": [["2000-01-01T00:00:00Z", "2024-01-01T00:00:00Z"]]},
+            "temporal": {
+                "interval": [["2000-01-01T00:00:00Z", "2024-01-01T00:00:00Z"]]
+            },
         },
         "links": [],
         "assets": {},
