+++ stac_fastapi/types/stac_fastapi/types/stac.py
@@ -2,6 +2,7 @@
 
 from typing import Any, Dict, List, Literal, Optional, Union
 
+import attr
 from stac_pydantic.shared import BBox
 from typing_extensions import TypedDict
 
@@ -74,3 +75,69 @@ class Collections(TypedDict, total=False):
 
     collections: List[Collection]
     links: List[Dict[str, Any]]
+
+
+class PartialCollection(TypedDict, total=False):
+    """Partial STAC Collection."""
+
+    type: Optional[str]
+    stac_version: Optional[str]
+    stac_extensions: Optional[List[str]]
+    id: Optional[str]
+    title: Optional[str]
+    description: Optional[str]
+    links: List[Dict[str, Any]]
+    keywords: Optional[List[str]]
+    license: Optional[str]
+    providers: Optional[List[Dict[str, Any]]]
+    extent: Optional[Dict[str, Any]]
+    summaries: Optional[Dict[str, Any]]
+    assets: Optional[Dict[str, Any]]
+
+
+class PartialItem(TypedDict, total=False):
+    """Partial STAC Item."""
+
+    type: Optional[Literal["Feature"]]
+    stac_version: Optional[str]
+    stac_extensions: Optional[List[str]]
+    id: Optional[str]
+    geometry: Optional[Dict[str, Any]]
+    bbox: Optional[BBox]
+    properties: Optional[Dict[str, Any]]
+    links: Optional[List[Dict[str, Any]]]
+    assets: Optional[Dict[str, Any]]
+    collection: Optional[str]
+
+
+@attr.s
+class PatchAddReplaceTest:
+    """Add, Replace or Test Operation."""
+
+    path: str = attr.ib()
+    op: Literal["add", "replace", "test"] = attr.ib()
+    value: Any = attr.ib()
+
+
+@attr.s
+class PatchRemove:
+    """Remove Operation."""
+
+    path: str = attr.ib()
+    op: Literal["remove"] = attr.ib()
+
+
+@attr.s
+class PatchMoveCopy:
+    """Move or Copy Operation."""
+
+    path: str = attr.ib()
+    op: Literal["move", "copy"] = attr.ib()
+
+    def __attrs_init__(self, *args, **kwargs):
+        """Init function to add 'from' field."""
+        super().__init__(*args, **kwargs)
+        self.__setattr__("from", kwargs["from"])
+
+
+PatchOperation = Union[PatchAddReplaceTest, PatchMoveCopy, PatchRemove]
